Provide default optimizer Adam to TrainingConfig
Switch from using rescaleGrad(1.0/batchSize) to having the loss use the mean
across the batch. This also requires updating some test cases and allows for the
batchAxis argument to be removed from the loss classes. Rescale grad is no
longer required and defaults to 1.0. This requirement for Loss classes is
documented in the Loss javadoc.

Create an (mostly placeholder) example to showcase various optimizers and how to
use them. Use the default optimizer to simplify the other examples.

Change-Id: I8024972449c4422db6ddecd820492914a1110f1a