Disentangle training examples from listener
Splits off the training listener so that each training example is self-contained
and all code is in external utilities. This both makes the examples easier to
understand and is progress towards moving those utilities out of the examples
and into the main api.

The trainer passes itself into the TrainListener methods
Trainer has a getMetrics method

Change-Id: I6c8d0f603475851f670868036a07a67359e2cb63