Remove the Matrix
Current there is no clear use case for this. Majority of the use cases could be achieved by NDArray.
NumPy initially bring this up for linear algebra but now they recommend to use NDArray instead and say
"The class may be removed in the future."  I think we can add it back if we see strong customer request.

Change-Id: I730d8f0281a2fa8eba6c08eaf1c3c2082124d6d4