Revert "Fix incorrect progress total in Batch"
This reverts commit fec7843db490a03f80a7a72da5168b9ccd729bc7.

The progress increments by the size of the batch, not by 1. So, the total should
be the total size of a dataset. This was chosen with consideration that we may
start having variable batch sizes where incrementing by 1 would not be as
informative. It should be correct with the exception of maxIteration, but that
is only intended for testing anyway.

Change-Id: I90daf00250ca4399c158ce75bce86860a258bb2f