Yolo GluonCV model
Create base ObjectDetectionTranslator

Create directory for the modality.cv.translator.wrapper to better organize

Create repository.getResourceDirectory(artifact) to handle the commonly repeated
code of finding the cached resource location and to handle it differently for
the simpleRepository

Allow an (optional) metadata file in the SimpleRepository to ease the process of
developing a model using it and testing with the metadata as well

Share the code for a python script that will import all of the yolo models into
the proper metadata format. It can be made more general to also apply for other
kinds of models, although it will probably have to be adjusted for each model
type.

Change-Id: I31fc5b002b39720ea94a4183c3bbf93d72df518c