Merge forward and predict into forward with boolean
The two methods seem to overlap most of the time, but it could be difficult to
handle the code duplication having them separate. I think passing the
information through the call stack as a boolean would be more intuitive and less
likely to be done wrong (especially with higher level blocks). For cases where
the training and inference passes differ significantly, they can just decompose
it into different methods.

Change-Id: Ie1f7ea27dc7264a8769767e2069bcce533b4a374