Update DataIterable to apply separate batchifier for data and label.
Dataset allows only one Batchifier to be set. While the implementation of
Batchifier can adjust for the various different kinds of inputs, there is
no way differentiate between data and labels since they share a Batchifier.

One clear example would be trying to train a sentiment analysis model for
StanfordMovieReview. The data is text data of varying length. We need to
use a PaddingStachBatchifier on the data due to the varying length. However,
the label is just a boolean that denotes true or false and just needs the
StackBatchifier.

Change-Id: Idf023fe656ca269f07115cae1d1979cebfc92069