Allow for full training outside of trainer experience
This moves the trainBatch and validateBatch methods to an external utility. This
ensures that a user can implement custom implementations of those methods while
still accessing all the features (like listeners) that the built-in ones take
advantage of.

The utility they are moved to is named EasyTrain. The previous TrainingUtils and
TrainingUtils.fit was combined into EasyTrain to make the full fit method
available to external users outside of the examples as well.

As part of this, I made model name required to better implement the training
listener and with the plan to later use that name to simplify the various usages
of model names.

Change-Id: Icc73489dda9265727d26433c7ae4463239fa0267