Use enhancement label for GitHub
We used to have both "enhancement" and "feature request". Instead, we should
just use one of them since they are roughly the same. I am choosing enhancement
because it is one of the default labels on github so it will be slightly more
familiar and work better with tools like the contribute page.

Change-Id: I360c5cd9479a5a7b0187cd66831744565afdeb6e