Refactor trackers to separate out warm up
Instead of warm-up being a core feature of trackers, it is a tracker itself that
composes with the other trackers. This simplifies the implementation.

Several of the trackers are also switched from using a step based update to a
pure function. The function should make it simpler to implement and calculate
the appropriate values to use for the trackers.

Change-Id: I6dc89c3436330db3858a9b8986f3ddaeba33df36