Rename the 3rdparty to extensions
This rename is clearer in purpose as it provides supplementary functionality on
top of the standard DJL API. Much of that functionality exists through
integration with additional libraries, but that still fits within the idea of an
extension.

The main improvement is that it enables us to add things which are first party
but still not considered for the main API. This might include a visualization
library or other things that might incidentally involve dependencies. It also
removes the numeral 3 from the module name which is somewhat messy.

Change-Id: I8909178551f511200aee5607e6804cab31bc203d