Include demos as part of main docs (#400)
Right now, the demo docs are a separate and easily missed section of the docs.
Instead, the demos can be combined into the main docs site. This means that we
only need to support a single site and users can find everything on that site
whether it is integrative or not. Also importantly, it unifies the docs search
between the two sites.

Change-Id: If1b262763fa69a5de0fd589113b6ca89ec5ba6b5