List input and output classes in ModelZoo lookup (#624)
* List input and output classes in ModelZoo lookup

This will list the input and output classes when trying to find a model in the
model zoo. It works either when no classes are specified or invalid classes are
specified. This will help users who don't know what the possible classes are to
resolve the question without looking at the code.

Change-Id: Ideecb0c0b56699a7023d50f2cf3da54010197112

* Fix

Change-Id: I64d4616f355aaaf13897aa44651867d8487c6ff9

* Use getFactoryLookupErrorMessage instead of printTranslatorFactories

Change-Id: I95719122be62952a6b9793571d345656eb7a99bd