Create ObjectDetectionDataset
This dataset simplifies the creation of new object detection datasets by
presenting a clearer and more standard class to extend.

This PR first creates an ImageDataset class that is the parent class of both the
ImageClassificationDataset and this ObjectDetectionDataset.

The ObjectDetectionDataset is then used as the new parent for the BananaDataset,
PikachuDataset, and CocoDetection dataset. They are all modified to use it.

In addition, some additional documentation is added to Rectangle to document
that it usually has coordinates from 0 to 1.

Change-Id: Ide12c593bcbc130ce574457dfacef6092cb111b2