Reduce the Conv testing during normal run (#810)
Currently, the convolution testing tests with a range of inputs across a number
of variables. This increases the test duration a bit much for the standard unit
test suite.

Instead, I reduce down to only two tests for unit testing. During nightly, the
full combination of ranges will be used.

Change-Id: I3b9671e2c353315368d5dff07271607cc6917c4c