Fixes TrainBertOnCode and temp attach chains (#786)
fixes #763

This first fixes the PointwiseFeedForwardBlock and simplifies it by having it
extend the SequentialBlock.

Then, it fixes an issue with the NDManager when a resource that was tempAttached
is then tempAttached again. So, the resource moves from managers A -> B -> C. It
was attached in A, got tempAttached to B, then tempAttached to C. When C was
closed, the resource was normal attached in B so it was closed too early. This
PR makes B remember detached temp resources, so even when it is returned from C
to B, B still knows that it should be tempAttached to return to A.

Change-Id: I4eae3415409e9908f21d9e7ec78996d66d3668a4