Converts Bert example to use Dataset class (#836)
* Converts Bert example to use Dataset class

Before, the example managed the data in a more ad-hoc fashion. While it still is
not a random access dataset, using the Dataset interface at least allows it to
integrate with EasyTrain.

This mostly splits the existing Dataset management code into the new
BertCodeDataset. The BertCodeDataset was not changed too much from it's original
design. Some simple cleanups were made to the remaining code in the
TrainBertOnCode example.

Change-Id: I7a150c213d59ca11d33e1560bba629a26d1bbde9

* Minor refactor

Change-Id: I2c37ddfd3311dc6d25ba96e11eb48a9140167441

Co-authored-by: Frank Liu <frankfliu2000@gmail.com>