BertCodeDataset uses Vocabulary and clean Vocabulary (#851)
* BertCodeDataset uses Vocabulary and clean Vocabulary

This refactor makes the BertCodeDataset use the standard vocabulary class
instead of it's custom Dictionary.

In addition, there are several improvements to the Vocabulary as well. The
SimpleVocabulary was renamed to DefaultVocabulary as it is not that simple. The
FtVocabulary was removed as the default can be used instead.

The pruning of the DefaultVocabulary was also improved. It adds the option to
limit the vocabulary to a fixed number of elements (by frequency) which is a
common usage in addition to requiring a minFrequency. This specifically was used
in the BertCodeDataset. It changes the default of the SimpleVocabulary as well
to not have a default minFrequency of 10 but to require an explicit one. I
believe this makes more sense as 10 is a very arbitrary number and users won't
expect their vocabularies to be randomly pruned.

Change-Id: I3a327322ba1ad42dad45bf1402825f58dfe01a21

* Fix comments

Change-Id: Icb32e3fc23adaefd8f4eb158642794154cf80a2b