Removes getPipeline from PreProcessor (#1245)
The pipeline is not a universal part of the PreProcessor and is rarely used. So,
it shouldn't be part of the main API but could be used by various
implementations.

This also cleans up the InstanceSegmentationTranslator a bit. It no longer
creates a new pipeline every time preprocess is called but uses a single
pipeline. It also no longer extends Transform but moves the necessary Transform
functionality to a nested class. Making it extend transform is odd and
potentially confusing, while the separation helps clarify the purpose of the
Transform.