Create TestRequirements utility
The TestRequirements simplifies the usage of SkipExceptions to skip various
tests. It helps reduce some code duplication of the requirements to concentrate
most of it within the same file. Finally, by aggregating the various
requirements into methods it makes it possible to use IDE findUsages to find all
of the tests with a particular requirement.