Create IndexEvaluator and IndexLoss (#1414)
Right now, there is inconsistent handling for how evaluators and losses can be
applied to only a single array within an NDList. This can happen often when
using multiple losses such as with the AbstractCompositeLoss cases.

Before, we had indexing through SimpleCompositeLoss. However, this would not
work with the trainer evaluators as well. So, the solution must be to move the
tracking of indices into the evaluators/losses.

This creates the indexed variations as a wrapper class so it requires less
updating for existing classes and keeps them simpler. As part of it, the classes
that already had indexing support including some accuracy evaluators and
SimpleCompositeLoss had it removed.

Change-Id: I9379ec32870210c116939e5c57a25aee72d5e356