Update the Dataset documentation
This makes a number of updates to the dataset documentation. It moves the list
of datasets to the main guide which is more accessible that basicdataset's README
and could contain future datasets not in basic dataset.

It also updates the list of built-in datasets and adds a list of dataset
helpers. It improves the page for creating a custom dataset. Finally, there are
a few other miscellaneous doc improvements.