Creates PreparedFeaturizer (#1700)
This creates a new concept for the PreparedFeaturizer which are used by
TabularDatasets. There are several kinds of Featurizers which require knowing
the possible inputs in advance such as the oneHot categorical encoding or a
normalized numeric encoder. That is what these are for.

There is also a breaking change. The default StringFeaturizer is changed from a
standard category index to a oneHot prepared featurizer. I believe that the
oneHot one tends to perform better during training, so I think it stands as a
better default now that it doesn't require extra work to use.

As part of this, it also moves several classes including DynamicBuffer, Feature,
Featurizer, and Featurizers from basicdataset/utils to
basicdataset/tabular/utils as they are more strongly associated with tabular
rather than general datasets.