Normalized Numeric Featurizer (#1701)
Earlier, the AirfoilRandomAccess had to use custom code to add normalizing
functionality and couldn't rely on the featurizers. This creates a new
normalizing featurizer using the new PreparedFeaturizer and brings it into line
with the standard tabular dataset.

The changes in testAirfoilRemotePreprocessing come from an update to the
normalization. Before, it was normalized by calculating using the size (limited
to top 1500 elements in the dataset for the test). Instead, it normalizes using
the full dataset availableSize(). I feel like this would make more sense to not
have your data preprocessing be different between a limited dataset and the full
one and would be unexpected for users if it were.