Split EasyTrain to add EasyTrainer
This creates a more extensible and non-final class EasyTrainer to contain
training functionality formerly in EasyTrain. There have been a few requests to
be able to modify EasyTrain and this should enable it.

EasyTrain is then left as a static wrapper around a shared instance of the
EasyTrain class. There shouldn't be any API changes required to use EasyTrain.