Fix ames and various tabular improvements
First for Ames, this fixes a bug (misclassified column type) to fix ames. It
also simplifies the Ames dataset by removing the featureToMap and replacing it
with the standard usage of the categorical Featurizer.

For tabular, it moves the validation function from the builder to the
constructor for tabular. This fixes the problem of how easy it is to forget the
validation when creating a dataset. As proof of this, several datasets were
missing the calls to validate. It adds two new tests for all features and no
features in Ames.