Begin publishing DJL Zero
As DJL Zero now features three applications, it probably makes sense to at least
make it available through Maven Central. This change begins publishing it. It
also improves the documentation and README to explain DJL Zero.

While it is added to the docs site, it is only added with a relatively low-key
location within the extensions section. In the future as DJL Zero is improved,
it may be worth giving it a more prominent place within the Home tab.