Ensure GradientCollector can clear gradients (#2101)
It also adds a new function GradientCollector.zeroGradients().

As part of this, it adds a new feature to get the arrays managed by an NDManager
or an NDResource. This is so all arrays can be found and have their gradients
cleared when the PtGradientCollector is started.

It also means that the System NDManagers are modified to now begin tracking
resources. As most things created under them are BaseNDManagers, it shouldn't be
a big issue. Then, operations that don't make sense under the SystemNDManagers
such as tempAttach or close have been changed to throw an exception rather than
work silently.