Test accumulating gradient collector
This adds additional documentation and testing to define the API for the
gradient collector. Specifically, it focuses on the gradient accumulating
behavior and adds a restriction to creating multiple simultaneous global
gradient collectors.

It also raises an issue for dealing with MXNet to accumulate gradients. I
investigated both the setRequiresGradient with ADD mode and backwards with
retainGrads and neither was correctly implementing the gradient accumulation
behavior.
As this is not a terribly significant use case, I will just leave it as a TODO
until there are any users who specifically request it.