Create translator options (#2145)
* Create translator options

This creates a new tool for TranslatorOptions based on the characteristic
function ((InputClass, OutputClass) -> Translator). Unlike the
TranslatorFactory, it doesn't require arguments or a model. So, the
TranslatorFactory is mainly used as part of model loading where the
TranslatorOptions can be more broad.

In particular, it connects to both Dataset and Translator. Both of them gain a
new function that will return a TranslatorOptions (matchingTranslatorOptions for
the Dataset and getExpansions for the translator). Right now, I made both of
these optional with a default value of null so that users don't have to define
them, it is just a possibility.

To connect the TranslatorOptions and TranslatorFactory, I created a class
ExpansionTranslatorFactory. This is used for the common abstraction where we
have a base translator and multiple "expansions" for it that come before it in
pre-processing or after it in post-processing. With a way to construct the base
translator with the options, it works as a TranslatorFactory. With a starting
Translator, it can become the TranslatorOptions.

Lastly and probably worth calling out, the ExpansionTranslatorFactory creates a
map of the options. This allows the supported types and the functions to be
defined in only one place instead of two. In a followup PR, I will make a
version of this that creates a BaseImageTranslatorFactory matching the
BaseImageTranslator and will handle all of the repeated code for all of the
image translator factories.

* minor javadoc updates

Co-authored-by: Frank Liu <frankfliu2000@gmail.com>