Improves TranslatorExpansions with pre-processing and post-processing
This provides an improvement to the TranslatorExpansions model. Now, it includes
three types of expansions: pre-processing expansions, post-processing
expansions, and full translator expansions. Before, it only had the full
translator expansions.

The key improvement is that any combination of pre-processing expansion and
post-processing expansion can be used. So, if there are 5 pre-processing and 3
post-processing then it will be 15 expansions. It also follows a standard
tendency to handle the pre-processing and post-processing separately as seen
with classes like the BaseImageTranslatorFactory.

Let me talk about the model for the expansion a bit. The translator can be
thought of as the composition of 5 different functions: pre-processing ->
batchify -> model -> unbatchify -> post-process.

With an expansion, it increases this to 7: pre-processing expansion ->
pre-processing -> batchify -> model -> unbatchify -> post-process ->
post-processing expansion.

The weakness of this implementation is that it also limits it to 7. For example,
we can use a Translator<InputStream -> Classifications> and expand the
pre-processing to form the Translator<Url -> Classifications>. With this model,
there is no built-in way to do this. That being said, this manifests mostly in
increased work on the side of writing expansions rather than a loss in
functionality. You would just have to use the InputStreamImagePreProcessor as
part of the UrlImagePreProcessor or otherwise write it from scratch. In exchange
for increased simplicity over a more complete model, I think it is a worthwhile
tradeoff.

Following this creation, there are a few other changes which I will list here
for ease of reading:
1. It creates a BasicTranslator which simply composes a
separate PreProcessor and PostProcessor. It is a useful utility in it's own
right
2. Moves the FileTranslator, UrlTranslator, and InputStreamTranslator into just
a PreProcessor
3. Uses the new ExpansionFactory to update the BaseImageTranslatorFactory and
apply it where possible