Creates the TabularTranslator
This creates a standard TabularTranslator based on the Featurizer and Feature
interface. Accompanying it is an ExpansionTranslatorFactory so it has two
possible inputs (MapFeatures or ListFeatures) and three outputs (TabularResults,
Regression, and Classification).

The MapFeatures and ListFeatures are created to help avoid complications with
class lookups. Otherwise, it would require weird unchecked casting:
(Class<List<String>>)(Class<?>)List.class.

As part of this, it also expands the Featurizer to also include capabilities for
deFeaturizing. The helper of DataFeaturizer is still around for Featurizers that
don't support the label deFeaturization.