Fixes PyTorch Java optimizers
PyTorch requires calling "no_autograd_guard{false}" in order to allow the
in-place operations required for an optimizer. This moves that call from each
individual optimizer into a single call at the end of the gradient collector.
This now makes it possible to implement a Java optimizer (like is done in the
D2l book).

To verify this, a new test is added that checks that a Java optimizer with
in-place operations will work.