Creates standard for PreTrained behavior
This changes the standard for DJL behavior with preTrained blocks. As of now,
they should also start out with frozen parameters. This has been applied to the embeddings.

It was previously applied only to PyTorch, but as of now applies to all models.
However, I did leave a carveout for models. It adds a boolean "wasLoaded" so
that if you load a model and then create a Trainer directly from it, it will not
be frozen. If you load a model and then append some new layers to it (as we have
several examples of), then it will need to be unfrozen.