Streaming Predict and streamable BytesSupplier
This is to replace #2447 to add support for an iterable streaming version of
predict. It creates a StreamingBlock and a streamingTranslator. Then, both are
added as part of a new call on Predictor.

In addition, it also adds two new BytesSupplier: a pull based
IterableBytesSupplier and a asynchronous Pub-Sub based PublisherBytesSupplier.